Make variable getting to-spec
It is expected by the UEFI specification that when a variable is
retrieved with NULL data, the data length is 0. Previous behaviour in
M2libc did not conform.

It appears some UEFI implementations depend on this and some do not
(notably, EDK2 does not, but many bare metal platforms do).