Fix UEFI writes acting strange after M2-Planet changes
https://github.com/oriansj/M2-Planet/pull/120/commits/413c69f4a58ff3e294572549875195df9ccfb5a8
introduced a regression for UEFI where output was not working correctly.
This is because UEFI uses UTF-16 for strings and in the commit we
allocated stack space by pushes which zeros the upper bytes while
store_value did not.
The UTF-16 string required 3 byte of zeroes after the initial character
value.
Because we only wrote to the lower part of the register the top parts
were effectively uninitialized.